Mapear relaciones con JPA y Hibernate

+++ aerolinea/src/main/java/co/edu/usbcali/aerolinea/dtos/AvionDTO.java
@@ -12,7 +12,7 @@ import java.util.Date;
 @Builder
 @ToString
 public class AvionDTO {
-    private Long avioId;
+    private Integer avioId;
     private String modelo;
     private String estado;
